requirements.txt
# Core dependencies
torch
torchvision
torchaudio
opencv-python
numpy<2.0.0
pandas
Pillow
tqdm

# Audio processing
ffmpeg-python
openai-whisper
librosa
soundfile
scipy

# Audio source separation
dora-search
pyyaml
demucs

# Computer vision and ML
ultralytics
transformers
scikit-learn

# Qwen multimodal model
modelscope
transformers_stream_generator
einops

# Model optimization
bitsandbytes
accelerate
auto-gptq

# Image processing
scikit-image

# Utilities
matplotlib
seaborn
jupyter
ipykernel
shazamio
protobuf


git+https://github.com/openai/CLIP.git

